+++ oidc-idp/src/main/java/cz/muni/ics/oidc/models/PerunAttributeValue.java
@@ -1,6 +1,5 @@
 package cz.muni.ics.oidc.models;
 
-import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.node.ArrayNode;